PoC streaming encoder for data frames
This adds a streaming encoder for data responses

It doesn't bound the size of frames, but does write out the frame in
in chunks.

This is the flip side of reading a large result set from a datasource
and creating frames without reading the whole result in to memory in one
go.

To test it locally, add this to `conf/custom.ini`:

    [feature_toggles]
    enable = jsonStreamingV2